API][CORE] SharedProjectMapper: Using IReferencePoint
SharedProjectMapper/-test:
This class now with reference points with resources now.

SarosSession:
After the SharedProjectManager works with reference points, the SarosSession needs
the ÍReferencePointManager for getting IProjects given by IReferencePoints.
The IReferencePointManager will instantiated in SarosSession and kept in
the core session container. The IReferencePointManager will needed only if a session is
running, so that the reason, that the IReferencePointManager is in the core session
container. Currently the SarosSession is responsible for putting the pairs IReferencePoints
and core.IProjects, which should to be shared.

CollaborationUtils/AddProjectsToSessionWizard:
The IReferencePointManager will filled after the session was started.