[FIX][I] #445 Avoid opening editor before sending file creation activity
Moves opening the created file in an editor after the file creation
activity was dispatched. This is necessary as the Saros otherwise
dispatches the editor opened activity before the file creation activity,
meaning the other participants receive an editor activation activity for
a non-existent resource. This is especially troublesome when the other
participant is following the user that created the file, as they then
automatically try to react to the editor activation activity by also
opening the (non-existent) file.

Fixes #445.