[REFACTOR][I] Do minor cleanup in LocalFilesystemModificationHandler
Moves the javadoc concerning the listener behavior from the handler
methods to the javadoc of the listener methods.

Fixes typo in method name 'generateResourceMoveActivity'.

Adds comment for IntelliJ code inspection to ignore the switch case in
generateRenamingResourceMoveActivity(...) having to few cases. The
switch gives us flexibility if we ever decide to add more cases and it
does not reduce the readability by much.